Separate out pvr_misc API.
Split from #539. Split out various misc
global and stats functions matching
`pvr_misc.c`
